+++ src/main/java/org/prgms/kdtspringweek1/controller/consoleController/ConsoleInputConverter.java
@@ -1,7 +1,7 @@
-package org.prgms.kdtspringweek1.consoleController;
+package org.prgms.kdtspringweek1.controller.consoleController;
 
 import org.prgms.kdtspringweek1.console.ConsoleInput;
-import org.prgms.kdtspringweek1.consoleController.dto.SelectFunctionTypeDto;
+import org.prgms.kdtspringweek1.controller.consoleController.dto.SelectFunctionTypeDto;
 import org.prgms.kdtspringweek1.voucher.service.dto.SelectVoucherTypeDto;
 import org.prgms.kdtspringweek1.voucher.service.dto.UpdateVoucherRequestDto;
 import org.prgms.kdtspringweek1.voucher.entity.VoucherType;
