+++ src/test/java/org/prgms/kdtspringweek1/controller/ConsoleInputConverterTest.java
@@ -7,9 +7,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.prgms.kdtspringweek1.console.ScannerInput;
-import org.prgms.kdtspringweek1.consoleController.ConsoleInputConverter;
+import org.prgms.kdtspringweek1.controller.consoleController.ConsoleInputConverter;
 import org.prgms.kdtspringweek1.voucher.service.dto.CreateVoucherRequestDto;
-import org.prgms.kdtspringweek1.consoleController.dto.SelectFunctionTypeDto;
+import org.prgms.kdtspringweek1.controller.consoleController.dto.SelectFunctionTypeDto;
 import org.prgms.kdtspringweek1.voucher.service.dto.SelectVoucherTypeDto;
 import org.prgms.kdtspringweek1.exception.InputExceptionCode;
 import org.prgms.kdtspringweek1.voucher.entity.VoucherType;
